React Profile Page setup
Created basic view and user interaction for sign in. Awaiting Profile route clean-up to include validation step to the sign in.
Currently, the form will require both a valid email and a password (can be anything that follows regex) to then redirect to profile.
Code is currently not commented, requires some refactoring and cleanup. If there is any question, ask me in discord.