Added comments to LogInPage
some minor comments to login page, logic need not be understood.
To test: nagivate to Login, then enter any email or password.
A success will redirect, a failure will maintain the page (no error return as of yet)
If successful, token will be saved. To clear the token for further testing, run
localStorage.clear()
in console to remove token.